refactor(payroll): lot 4 — paiements : CreatePaymentBatch + ConfirmPaymentItemReception extraites de PaymentBatchController, contrôleur allégé (Part of #6968, ADR-0020)
