+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - **fix(ci/deploy): rattrapage du tier dev débloqué — input \`workflow_dispatch.force_deploy\` (Closes #6973, réouverte).** Cause (logs des runs) : (1) sous rafale de merges, la gate interne trouve la conclusion du run Tests du SHA = \`cancelled\` → \`should_deploy=false\` → job deploy **skippé** (aucun deploy exécuté depuis le 2026-09-07 16:07, SHA live figé \`623ed31\`) ; (2) le rattrapage horaire #6984 re-dispatchait sans pouvoir forcer → un HEAD de main sans diff \`api/\` (commit .github/docs) rend \`api_changed=false\` → skip en boucle chaque heure. Correctif : \`deploy-main.yml\` accepte désormais l'input \`force_deploy\` (workflow_dispatch uniquement, chemin push inchangé) et \`deploy-main-catchup.yml\` le passe à \`true\` — le HEAD de main est déployé (Render construit le HEAD : le backlog api est livré d'un coup ; le code a déjà passé les checks requis de la PR avant merge).
 - **fix(ci/ops): deploy-main — le gate vérifie désormais le SHA réellement servi par le tier dev (Closes #6834, #6973).** Constat 2026-09-06/07 : `deploy-main.yml` « réussissait » depuis ~24 h sans que le service Render dev (`gestionemployerbackend`) n'applique main (dernier sha live figé `9aca97a`, routes web DEV en 500 #6957, provisioning trial en échec #6958). Causes : le healthcheck post-hook (1) pouvait cibler une autre instance via le secret partagé `API_HEALTHCHECK_URL` et (2) exigeait `version == APP_VERSION` (4.24.0) alors que le tier dev n'a pas d'`APP_VERSION` : `/api/v1/health` y expose le SHA Render tronqué (`RENDER_GIT_COMMIT`, résolution « version honnête » #6835/#6905) → comparaison structurellement impossible, gate aveugle. Correctif : `version` baseline capturée AVANT le hook (`LIVE_BASELINE_VERSION`) ; le job d'attente exige une instance NOUVELLE (version ≠ baseline) servant le HEAD courant de main (sha7 d'`origin/main`, robuste sous rafale de merges) ou le SHA déclencheur exact (re-run d'un SHA déjà live) ; URL de santé verrouillée sur l'instance déployée (`vars.PROD_API_BASE_URL`, dette de nommage #6839) ; échec = sortie rouge explicite + résumé step. Ajout de `deploy-main-catchup.yml` (horaire) : compare le SHA live au HEAD de main et redéclenche `deploy-main` si retard (gate interne : Tests du HEAD verts, aucun run déjà en cours) — un run skippé ne laisse plus le tier dev en retard sans signal rouge. Aucun changement applicatif.
 
+- **refactor(payroll): couche Application — lot 4 paiements, 2 Actions extraites de `PaymentBatchController` (Part of #6968, ADR-0020).** (1) `CreatePaymentBatch` (Application/Actions) : création d'un lot de paiement (draft) + ses `PaymentItem` `pending` depuis un run calculé/validé/payé — résolution du run bornée au tenant, gardes métier (statut du run `PAYMENT_BATCH_RUN_INVALID`, lot sans bulletin payable refusé) et transaction de création ; (2) `ConfirmPaymentItemReception` (Application/Actions) : confirmation de réception d'un paiement par l'employé destinataire — garde d'état (`pending` non confirmable, déclaration manager requise), création **idempotente** de la `PaymentConfirmation`, signature de consentement horodatée PA2-PAY-016 (hash calculé sur le `confirmed_at` réellement persisté après `refresh`, jamais sur la valeur non arrondie), transition du `PaymentItem` vers `confirmed` et rafraîchissement du statut agrégé du batch (`confirmed` / `partially_confirmed`). `PaymentBatchController` allégé (337 → 226 l.) : gardes HTTP (404 company/employee), validation de requête et enveloppes de réponse conservées au niveau interface ; `ip`/`user_agent` passés en paramètres de l'Action. **Aucun changement de comportement API** (mêmes routes, statuts, messages et écritures d'audit — refactor sous tests `PaymentBatchControllerTest`). Vérifié : php -l OK, Pint 1.30.5 OK.
 - **fix(ci/deploy): garde de déploiement Render alignée sur la version honnête — SHA git accepté (Part of #6973).** Depuis #6905/#6946, `/api/v1/health` renvoie `config('app.version')` = APP_VERSION si l'env Render la pose, sinon le **SHA git court** (fallback `RENDER_GIT_COMMIT`, 7 car.) — mais la garde « Wait for API/Web healthcheck » de `deploy-main.yml` comparait encore à PROGRAM_VERSION (4.24.0) → **faux échecs systématiques** (« did not detect the new instance » après 60×20 s + rollback) alors que Render avait bien déployé le nouveau SHA (observé sur les runs 623ed31/aab24dd6, 2026-09-07/08). La garde accepte désormais le SHA court du commit déployé (`DEPLOY_SHA[:7]`) OU l'APP_VERSION, et loggue clairement l'attente. Aucun changement de la logique uptime/cold-start (requise inchangée) ni du déclenchement.
 - **docs(architecture): BC-27 spike éditeur — décision : éditeur de sections MAISON en v1, GrapeJS écarté (Closes #6869, V-SPIKE).** Rapport `docs/architecture/SHOWCASE_EDITOR_SPIKE_2026-09-08.md` : GrapeJS (et les builders headless) a un modèle de document HTML libre orthogonal au contrat de sections typées #6866 (double modèle de contenu, mapping inverse fragile, dépossession du rendu des thèmes serveur #6868, sanitisation HTML + CSP, i18n/RTL hors modèle) — écarté pour le v1 (7 types, 3 thèmes). V1 = éditeur maison (liste de sections + formulaires par type + aperçu), zéro dépendance ajoutée à l'admin-dashboard ; déclencheurs de ré-évaluation documentés (sections libres, WYSIWYG, vitrine généraliste). V-EDITOR #6870 cadré en conséquence. Aucun changement de code applicatif.
 - **feat(ai): A5 — audit conversation→action→effet : journal d'exécution d'outils `ai_tool_executions` avec marqueur « via assistant IA » (Closes #6852, BC-23, EPIC #6846).** Chaque étape de la chaîne est tracée de façon rejouable : (1) outil de LECTURE exécuté pendant le chat → ligne `stage=executed` reliée à la conversation ; (2) écriture sensible → ligne de proposition `stage=confirmation_required` avec `pending_action_id` ; (3) confirmation humaine → ligne d'exécution reliée au MÊME `pending_action_id` (la proposition porte la `conversation_id`, la chaîne se reconstitue par jointure) ; (4) refus humain → `stage=rejected`, aucun effet métier. Colonnes : `tool_name`, `tool_input` (JSON SANITISÉ — PII masquées via PrivacySanitizer, A6 #6853), `stage`, `success`, `result_summary`, `source='assistant'`, horodatage serveur ; index (company_id, created_at), conversation_id, pending_action_id. Câblage : `AIAuditLogger::logToolExecution()` (nouveau), boucle d'outils de l'`Orchestrator` (chat), `AIGatewayController::confirmAction`/`rejectAction`. Migration tenant idempotente + miroir fixture `CreatesMvpSchema` (garde parity #5443). Tests Feature `AIToolExecutionAuditTest` : lecture journalisée avec conversation, chaîne écriture proposition→exécution par pending_action_id, refus sans effet. Vérifié : 3 tests verts (PHP 8.4 + PG16), Pint OK, php -l OK. Aucun changement de contrat API (routes/verbatim inchangés, écritures additives).
